po 20151205，修改：添加一个UIView统一布局labels，不再单个单个在cell中添加标签； po 20151205，遗留问题：如果直接固定tableview的高度，则tableview里面cell很多的时候，最下面的看不到；如果不固定，则会在操作的时候很容易进到tableview里面，无法返回detailview po 20151206，完成了detail view里面的所有布局，在basic cell中添加了edit按钮push edit view

+++ XF AppDemo V0.0/XFWork/CommonItem/Model/XFNWorkTableViewCellModel.m
@@ -35,7 +35,7 @@
     model.detailString    = [[NSString alloc] initWithFormat : @"%@平米 | %@房%@卫 | %@元/平米", [_assetAllInfoModel objectForKey : @"assetTotalArea"],
                                                                                                [_assetAllInfoModel objectForKey : @"quantityOfRoom"],
                                                                                                [_assetAllInfoModel objectForKey : @"quantityOfToilet"],
-                                                                                               [NSString stringWithFormat: @"%ld", unitPrice]];
+                                                                                               [NSString stringWithFormat: @"%ld", (long)unitPrice]];
     model.priceString     = [[NSString alloc] initWithFormat : @"%@万", tempPrice];
     model.statusString    = [_assetAllInfoModel objectForKey : @"assetStatus"];
     model.ownernameString = [_assetAllInfoModel objectForKey : @"attributeTo"];
